% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{AAVY}
\alias{AAVY}
\title{The function used for Average Annual Variability in Yield being less than 20\% Performance metric}
\usage{
AAVY(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for Average Annual Variability in Yield being less than 20\% Performance metric
}
