% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-kobe.R
\name{calc_contour_lines}
\alias{calc_contour_lines}
\title{Calculate management procedure contour lines for two quantities}
\usage{
calc_contour_lines(d, alpha = c(0.025, 0.5, 0.975), n = 200)
}
\arguments{
\item{d}{A list of management procedure data frames with columns mp, mp_name, x, and y.}

\item{alpha}{A vector of levels between 0 and 1 for the contour lines.}

\item{n}{As defined in \code{\link[MASS:kde2d]{MASS::kde2d()}}.}
}
\value{
A data frame containing mp, mp_name, alpha, x, and y where x and y are the calculated
coordinates for the contour lines for each alpha and mp.
}
\description{
Calculate management procedure contour lines for two quantities
}
