% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-rmd.R
\name{change_chunk_suffix}
\alias{change_chunk_suffix}
\title{Change the chunk and tag suffixes for a .Rmd file}
\usage{
change_chunk_suffix(file_name, chunk_suffix = "")
}
\arguments{
\item{file_name}{Filename/path of the .rmd file to create/modify. If it does not exist,
an error will be given}

\item{chunk_suffix}{A string to be appended to the chunk names and tags with a preceding dash.
If a name has already been appended this new suffix will replace it. If this is the empty
string, any previous suffixes will be removed.}
}
\value{
Nothing
}
\description{
Change the chunk and tag suffixes for a .Rmd file
}
\examples{
unlink("test.Rmd")
create_default_rmd("test.Rmd")
change_chunk_suffix("test.Rmd", "test-suffix")
unlink("test.Rmd")
}
