% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-rmd.R
\name{check_tags}
\alias{check_tags}
\title{Check that begin and end slot-chunk tags match properly}
\usage{
check_tags(rmd)
}
\arguments{
\item{rmd}{The lines of the Rmd file as read in by readLines()}
}
\value{
Nothing
}
\description{
Check that begin and end slot-chunk tags match properly
}
