% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-rmd.R
\name{create_default_rmd}
\alias{create_default_rmd}
\title{Create template Rmd file describing DLMtool Objects and Slots}
\usage{
create_default_rmd(file_name, overwrite = FALSE, knitr_results = TRUE,
  knitr_echo = TRUE)
}
\arguments{
\item{file_name}{Filename/path of where to save the .Rmd file.}

\item{overwrite}{Overwrite?}

\item{knitr_results}{Show knitr results?}

\item{knitr_echo}{Echo knitr code?}
}
\value{
Nothing
}
\description{
Create template Rmd file describing DLMtool Objects and Slots
}
\examples{
unlink("test.Rmd")
create_default_rmd("test.Rmd")
}
