% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-rmd.R
\name{create_rmd}
\alias{create_rmd}
\title{Create .Rmd file describing MSEtool Objects and Slots and inject custom descriptions into it}
\usage{
create_rmd(
  file_name,
  cust_desc_file_name = system.file("alt-slot-descriptions.csv", package = "ggmse"),
  slot_type_order_file_name = system.file("slot-type-order.csv", package = "ggmse"),
  ...
)
}
\arguments{
\item{file_name}{Filename of the .rmd file to create/modify. If it does not exist,
it will be created using \code{\link[=create_default_rmd]{create_default_rmd()}}}

\item{cust_desc_file_name}{Filename of the .csv file containing the custom descriptions.
Use generate_default_custom_descriptions_file() in scratch.R to auto-generate it.}

\item{slot_type_order_file_name}{Filename of the .csv file containing the slot orders.}

\item{...}{Arguments to be passed to \code{\link[=create_default_rmd]{create_default_rmd()}}}
}
\value{
Nothing
}
\description{
Create .Rmd file describing MSEtool Objects and Slots and inject custom descriptions into it
}
\examples{
create_rmd("test.Rmd")
}
