% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-om.R
\name{download_om}
\alias{download_om}
\title{Download an operating model from the DLMtool website}
\usage{
download_om(
  file = "default-download-om.rds",
  web_file,
  base_url = "http://www.datalimitedtoolkit.org/Case_Studies_Table/",
  method = "wget",
  overwrite = FALSE
)
}
\arguments{
\item{file}{The name of the file. If there is no .rds extension one will be appended}

\item{web_file}{Like this: "Pacific_Ocean_Perch_QC_BC_DFO/OM.rdata" See webpage for listing}

\item{base_url}{The URL to find the web_file on}

\item{method}{Download method. It seems that on "wget" works on MS Windows}

\item{overwrite}{TRUE or FALSE to overwrite file}
}
\description{
Download an operating model from the DLMtool website
}
\examples{
\donttest{
library(gfdlm)
download_om("pop", "Pacific_Ocean_Perch_QC_BC_DFO/OM.rdata")
}
}
