% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{filter_data_rex}
\alias{filter_data_rex}
\title{Filter Rex sole data}
\usage{
filter_data_rex(d, minimum_year = 1996, maximum_year = 2018,
  major_stat_area_codes = c("03", "04"),
  survey_series_descs = c("West Coast Vancouver Island Synoptic Bottom Trawl"))
}
\arguments{
\item{d}{data as retrieved from \code{\link[gfdata:get_commercial_samples]{gfdata::get_commercial_samples()}} and similar functions}

\item{minimum_year}{Minimum year for filtering}

\item{maximum_year}{Maximum year for filtering}

\item{major_stat_area_codes}{vector of major stat area codes}

\item{survey_series_descs}{vector of survey descriptions as they appear in the survey_series_desc
field of the survey_index item of the call to \code{\link[gfdata:get_survey_index]{gfdata::get_survey_index()}}}
}
\value{
The filtered data
}
\description{
Filter Rex sole data
}
