% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-summary.R
\name{FSP_summary}
\alias{FSP_summary}
\title{Summary of status estimates}
\usage{
FSP_summary(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  MPD = FALSE,
  LRP = 0.4,
  USR = 0.8,
  FUN = median,
  ref = rep(TRUE, length(rcm))
)
}
\arguments{
\item{rcm}{A list containing RCModel objects.}

\item{scenario}{A character vector of names corresponding to rcm.}

\item{MPD}{Logical, whether to plot individual simulations (FALSE) or from the single fit (TRUE) in RCModel@mean_fit.}

\item{LRP}{The fraction of BMSY for the limit reference point}

\item{USR}{The fraction of BMSY for the upper stock reference}

\item{FUN}{Function for summarizing results across simulations for each scenario, e.g., median or mean. Set to NULL to
return values from all simulations.}

\item{ref}{A logical vector to identify reference operating models for the summary statistics}
}
\value{
A data frame with:
\itemize{
\item B0
\item R0
\item h
\item M
\item Bt (terminal year spawning biomass)
\item Ft (terminal year fishing mortality)
\item BMSY
\item FMSY
\item MSY
\item B_BMSY
\item p_LRP (probability above the LRP)
\item p_USR (probability above USR)
\item p_BMSY (probability above BMSY)
\item F_FMSY
\item p_FMSY (probability below FMSY)
}
}
\description{
A wrapper function that returns biomass and F estimates for a set of conditioned operating models
to support reporting for the Fish Stocks Provisions.
}
