% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{IDX}
\alias{IDX}
\alias{IDX_smooth}
\title{Index-based MP from Cox et al. (2019)}
\usage{
IDX(
  x,
  Data,
  reps = 100,
  delta_min = -0.5,
  delta_max = 0.25,
  lambda = 1,
  tac_floor = NULL,
  year_ref = 1
)

IDX_smooth(x, Data, reps = 100, tac_floor = 1, ...)
}
\arguments{
\item{x}{A position in the data object.}

\item{Data}{A data object.}

\item{reps}{The number of stochastic samples of the MP recommendation(s).}

\item{delta_min}{Most negative drop proportion allowed in the index.}

\item{delta_max}{Most positive increased proportion allowed in the index.}

\item{lambda}{Smoothing parameter. 0 means always use the last TAC. 1 means
no smoothing. Can take any value in between.}

\item{tac_floor}{TAC when \code{delta_min} is met or exceeded. If left as \code{NULL},
the floor will be set to 20\% of average catch from the last
5 years.}

\item{year_ref}{Number of years before the present year for the comparison of
the index value. Default looks back one year. For a biennial survey this
should be changed to something larger such as 2 or the recommended TAC will
never change. Note that the algorithm fills in \code{NA} values in the index as
the previous non-NA value.}

\item{...}{Other arguments to pass to \link{IDX}.}
}
\description{
Index-based MP used in the most recent outside Yelloweye Rockfish assessment
in BC.
}
\examples{
IDX(1, MSEtool::SimulatedData)
IDX(1, MSEtool::SimulatedData, lambda = 0.5)
}
\references{
Sean P Cox, Beau Doherty, Ashleen J Benson, Samuel DN Johnson, and Dana
Haggarty. Evaluation of potential rebuilding strategies for Outside Yelloweye
Rockfish in British Columbia. Working Paper 2017GRF02.
}
