% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{Itarget}
\alias{Itarget}
\alias{Itarget_base}
\alias{Itarget_w0.8}
\alias{Itarget_x0.2}
\alias{Itarget_x0.8}
\alias{Itarget_d1.2}
\alias{Itarget_d0.8}
\title{Itarget MP}
\usage{
Itarget(
  x,
  Data,
  reps = 1,
  w = 0,
  lambda = 0.2,
  delta = 1,
  xx = 1,
  index_target_window = 10,
  index_current_window = 5,
  catch_target_window = 5
)

Itarget_base(x, Data, reps = 1)

Itarget_w0.8(x, Data, reps = 1)

Itarget_x0.2(x, Data, reps = 1)

Itarget_x0.8(x, Data, reps = 1)

Itarget_d1.2(x, Data, reps = 1)

Itarget_d0.8(x, Data, reps = 1)
}
\arguments{
\item{x}{A position in the data object. As per \pkg{MSEtool}.}

\item{Data}{A data object. As per \pkg{MSEtool}.}

\item{reps}{The number of stochastic samples of the MP recommendation(s). As
per \pkg{MSEtool}.}

\item{w}{A smoothing parameter that defines the "steepness" of the adjustment
slope.}

\item{lambda}{Fraction of the average index over the 10 years prior to the
projection period below which future TACs are reduced quadratically.}

\item{delta}{\verb{I_\{target\}} is \verb{(1 + \\delta) I_\{ave\}}; the target index value.}

\item{xx}{Catch target defined as \verb{(1 - xx)C_\{ave\}}, where \code{xx} is the
proportional difference between the future catch and the average historical
catch over the last 5 years of the historical period.}

\item{index_target_window}{Index target window in years.}

\item{index_current_window}{Index current window in years. Years over
which the average index is calculated.}

\item{catch_target_window}{Catch target window in years.}
}
\description{
Itarget MP
}
\examples{
Itarget(1, MSEtool::SimulatedData)
}
