% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{Itarget5}
\alias{Itarget5}
\title{Itarget 5 MP}
\usage{
Itarget5(x, Data, reps = 1, yrsmth = 5, xx = 0, Imulti = 2)
}
\arguments{
\item{x}{A position in the data object. As per \pkg{DLMtool}.}

\item{Data}{A data object. As per \pkg{DLMtool}.}

\item{reps}{The number of stochastic samples of the MP recommendation(s). As
per \pkg{DLMtool}.}

\item{yrsmth}{Years over which the average index is calculated.}

\item{xx}{Parameter controlling the fraction of mean catch to start using
in first year.}

\item{Imulti}{Parameter controlling how much larger target CPUE / index is
compared with recent levels.}
}
\description{
Itarget 5 MP
}
