% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{ITM_hist}
\alias{ITM_hist}
\title{Historical Index Target based on natural mortality rate}
\usage{
ITM_hist(x, Data, reps = 100, yrsmth_hist = 10, ...)
}
\arguments{
\item{x}{A position in the data object. As per \pkg{MSEtool}.}

\item{Data}{A data object. As per \pkg{MSEtool}.}

\item{reps}{The number of stochastic samples of the MP recommendation(s). As
per \pkg{MSEtool}.}

\item{yrsmth_hist}{Number of historical years to consider for reference index
level.}

\item{...}{Other arguments to pass to the MP function.}
}
\description{
This MP is based on \code{\link[DLMtool:ITM]{DLMtool::ITM()}} but since the reference index level to
the index over some historical time period.
}
