% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_data_factory}
\alias{load_data_factory}
\title{Factory to create loading functions for various species}
\usage{
load_data_factory(species_name, file)
}
\arguments{
\item{species_name}{The species name as can be used in extraction with \link[gfdata:get_commercial_samples]{gfdata::get_commercial_samples}
and similar functions}

\item{file}{The filename to load from}
}
\value{
A function which can be used to load data automatically
}
\description{
Factory to create loading functions for various species
}
\examples{
load_data_pcod <- load_data_factory("pacific cod",
  file = here::here("generated-data",
                    "pacific-cod.rds"))
}
