% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_data_rex}
\alias{load_data_rex}
\title{Loads Rex Sole data}
\usage{
load_data_rex(unsorted_only = TRUE, private = FALSE)
}
\arguments{
\item{unsorted_only}{If TRUE, return only unsorted records in the commercial samples}

\item{private}{If TRUE, do not include commercial samples and catch in the returned data}
}
\value{
A list with the outputs of \code{\link[gfdata:get_catch]{gfdata::get_catch()}}, \code{\link[gfdata:get_survey_samples]{gfdata::get_survey_samples()}}, \code{\link[gfdata:get_survey_index]{gfdata::get_survey_index()}}, and \code{\link[gfdata:get_commercial_samples]{gfdata::get_commercial_samples()}}.
}
\description{
Loads Rex Sole data
}
