% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{LTY}
\alias{LTY}
\title{The function used for Long Term Yield Performance metric}
\usage{
LTY(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for Long Term Yield Performance metric
}
