% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-all.R
\name{make_typical_plots}
\alias{make_typical_plots}
\title{Make all typical plots}
\usage{
make_typical_plots(
  mse_list,
  pm,
  scenario_df,
  this_year,
  mp_sat,
  mp_not_sat,
  mp_not_sat_highlight,
  mp_ref,
  custom_pal,
  eg_scenario,
  catch_breaks = NULL,
  catch_labels = catch_breaks,
  satisficed_criteria = NULL
)
}
\arguments{
\item{mse_list}{A list of DLMtool MSE objects representing different
scenarios. The list should be named with the scenario names.}

\item{pm}{A character vector of performance metrics. These performance
metrics should exist in current workspace.}

\item{scenario_df}{A data frame with the columns \code{scenario},
\code{scenarios_human}, and \code{scenario_type}. \code{scenario_type} should be one of
\code{Reference} or \code{Robustness}.}

\item{this_year}{The last year of historical data in the MSE.}

\item{mp_sat}{A character vector of satisficed to management procedures
(MPs).}

\item{mp_not_sat}{MPs that were \emph{not} satisfied (a projection plot will be
made with these).}

\item{mp_not_sat_highlight}{MPs that were \emph{not} satisfied to highlight
in a projection plot. I.e. some subset of the full not satisfied set.}

\item{mp_ref}{Reference MPs}

\item{custom_pal}{A named character vector of colors for the MPs.
Names should correspond to the MP names. Should include the
satisficed and reference MPs.}

\item{eg_scenario}{An example scenario (as character) which will be used for
the projection plot of not-satisficed MPs.}

\item{catch_breaks}{An optional numeric vector of y-axis breaks for the catch
projection panels.}

\item{catch_labels}{An optional numeric vector of y-axis labels for the catch
projection panels. This can be useful, for example, if you want the labels
to be in 1000 t insead of t.}

\item{satisficed_criteria}{A named numeric vector designating the satisficed
criteria for use in a tigure plot. See \code{\link[=plot_tigure]{plot_tigure()}}.}
}
\value{
A named list object containing the ggplot objects.
}
\description{
This function returns a large set of typical plots in a list object. These
can then be printed to a file or included in an R Markdown document.
}
