% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{.DD}
\alias{.DD}
\alias{.DD4010}
\alias{.DDSS_MSY}
\alias{.DDSS_4010}
\alias{.GB_slope}
\alias{.ICI}
\alias{.ICI2}
\alias{.Iratio}
\alias{.Islope1}
\alias{.Islope2}
\alias{.Islope3}
\alias{.Islope4}
\alias{.IT5}
\alias{.IT10}
\alias{.Itarget1}
\alias{.Itarget2}
\alias{.Itarget3}
\alias{.Itarget4}
\alias{.ITM}
\alias{.LstepCC1}
\alias{.LstepCC2}
\alias{.LstepCC3}
\alias{.LstepCC4}
\alias{.Ltarget1}
\alias{.Ltarget2}
\alias{.Ltarget3}
\alias{.Ltarget4}
\alias{.Ltarget95}
\alias{.SBT1}
\alias{.SBT2}
\alias{.SCA_4010}
\alias{.SCA_MSY}
\alias{.SP_4010}
\alias{.SP_MSY}
\alias{.SPMSY}
\alias{.SPmod}
\title{Biennial survey versions of many \pkg{DLMtool} and \pkg{MSEtool} MP functions}
\usage{
.DD(x, Data, reps = 100, ...)

.DD4010(x, Data, reps = 100, ...)

.DDSS_MSY(x, Data, reps = 100, ...)

.DDSS_4010(x, Data, reps = 100, ...)

.GB_slope(x, Data, reps = 100, ...)

.ICI(x, Data, reps = 100, ...)

.ICI2(x, Data, reps = 100, ...)

.Iratio(x, Data, reps = 100, ...)

.Islope1(x, Data, reps = 100, ...)

.Islope2(x, Data, reps = 100, ...)

.Islope3(x, Data, reps = 100, ...)

.Islope4(x, Data, reps = 100, ...)

.IT5(x, Data, reps = 100, ...)

.IT10(x, Data, reps = 100, ...)

.Itarget1(x, Data, reps = 100, ...)

.Itarget2(x, Data, reps = 100, ...)

.Itarget3(x, Data, reps = 100, ...)

.Itarget4(x, Data, reps = 100, ...)

.ITM(x, Data, reps = 100, ...)

.LstepCC1(x, Data, reps = 100, ...)

.LstepCC2(x, Data, reps = 100, ...)

.LstepCC3(x, Data, reps = 100, ...)

.LstepCC4(x, Data, reps = 100, ...)

.Ltarget1(x, Data, reps = 100, ...)

.Ltarget2(x, Data, reps = 100, ...)

.Ltarget3(x, Data, reps = 100, ...)

.Ltarget4(x, Data, reps = 100, ...)

.Ltarget95(x, Data, reps = 100, ...)

.SBT1(x, Data, reps = 100, ...)

.SBT2(x, Data, reps = 100, ...)

.SCA_4010(x, Data, reps = 100, ...)

.SCA_MSY(x, Data, reps = 100, ...)

.SP_4010(x, Data, reps = 100, ...)

.SP_MSY(x, Data, reps = 100, ...)

.SPMSY(x, Data, reps = 100, ...)

.SPmod(x, Data, reps = 100, ...)
}
\arguments{
\item{x}{A position in the data object. As per \pkg{DLMtool}.}

\item{Data}{A data object. As per \pkg{DLMtool}.}

\item{reps}{The number of stochastic samples of the MP recommendation(s). As
per \pkg{DLMtool}.}

\item{...}{Other arguments to pass to the MP function.}
}
\description{
These management procedures (MPs) use \code{\link[=reduce_survey]{reduce_survey()}} to eliminate odd
years of survey observations to reflect the biennial nature of most
groundfish surveys in British Columbia.
}
