% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\docType{data}
\name{mse}
\alias{mse}
\title{Example simple MSE}
\format{An object of class \code{MSE} of length 1.}
\usage{
mse
}
\description{
An object of class \linkS4class{MSE} with the MP set to "AvC"
}
\keyword{datasets}
