% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\docType{data}
\name{mse_example}
\alias{mse_example}
\title{Example simple MSE}
\format{
An object of class \code{MSE} of length 1.
}
\usage{
mse_example
}
\description{
An object of class \linkS4class{MSE} for examples
}
\keyword{datasets}
