% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om.R
\docType{data}
\name{om}
\alias{om}
\title{Example simple operating model}
\format{
An object of class \code{OM} of length 1.
}
\usage{
om
}
\description{
An object of class \linkS4class{OM} with nsim set to 3
}
\keyword{datasets}
