% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{P10}
\alias{P10}
\title{The function used for 10\% BMSY Performance metric}
\usage{
P10(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for 10\% BMSY Performance metric
}
