% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{P100_yrs21_35}
\alias{P100_yrs21_35}
\title{The function used for 100\% BMSY Performance metric for years 21-35}
\usage{
P100_yrs21_35(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for 100\% BMSY Performance metric for years 21-35
}
