% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{P80}
\alias{P80}
\title{The function used for 80\% BMSY Performance metric}
\usage{
P80(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for 80\% BMSY Performance metric
}
