% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{P80_yrs36_50}
\alias{P80_yrs36_50}
\title{The function used for 80\% BMSY Performance metric for years 36-50}
\usage{
P80_yrs36_50(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for 80\% BMSY Performance metric for years 36-50
}
