% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-convergence.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot convergence}
\usage{
plot_convergence(
  object_list,
  pm_list = c("LTY", "PNOF"),
  label_gap = 1.15,
  custom_pal = NULL,
  ylim = c(min(df$value), 1),
  satisficed = NULL,
  french = FALSE
)
}
\arguments{
\item{object_list}{A (named) list of DLMtool MSE objects.}

\item{pm_list}{A list or vector of performance metrics as character.}

\item{label_gap}{Controls the gap on the right of each panel to make space
for the MP labels. This value is multiplied by the maximum number of
iterations to expand the x axis.}

\item{custom_pal}{An optional custom color palette. Should be a named
character vector}

\item{ylim}{Y limits. Defaults to the minimum observed performance metric
value (via lazy evaluation) and 1.}

\item{satisficed}{An optional named numeric vector. The names correspond to
the performance metrics and the values correspond to the threshold.
This will add a horizontal line on the relevant panels.}

\item{french}{French?}
}
\value{
A ggplot2 plot.
}
\description{
This function plots cumulative performance metric values across iterations to
assess convergence.
}
\examples{
plot_convergence(mse_example)
plot_convergence(mse_example, satisficed = c("LTY" = 0.9))
}
