% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-dots.R
\name{plot_dots}
\alias{plot_dots}
\title{Dot plot}
\usage{
plot_dots(
  pm_df_list,
  type = c("single", "facet"),
  custom_pal = NULL,
  dodge = 0.6,
  bar_alpha = 0.2,
  pt_size = 2.25
)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames from
\code{\link[=get_probs]{get_probs()}}. The names will be used as the plot labels.
Can also be a single data frame from \code{\link[=get_probs]{get_probs()}}.}

\item{type}{The type of plot. Multipanel \code{"facet"} vs. single panel
\code{"single"}. In the single panel version, a line segment represents the
upper and lower values across the scenarios and the dot represents the
mean.}

\item{custom_pal}{An optional custom color palette. Should be a named
character vector}

\item{dodge}{The amount to separate or "dodge" the dots.}

\item{bar_alpha}{Background bar transparency. 0 will omit.}

\item{pt_size}{Point size.}
}
\value{
A ggplot2 object
}
\description{
Dot plot
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm <- list()
pm[[1]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm[[2]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
names(pm) <- c("Scenario 1", "Scenario 2")
plot_dots(pm)
plot_dots(pm, type = "facet")
}
