% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-radar.R
\name{plot_grid_pbs}
\alias{plot_grid_pbs}
\title{Plot a grid of ggplots}
\usage{
plot_grid_pbs(
  plotlist,
  align = "hv",
  label_fontface = "bold",
  label_size = 12,
  hjust = 0,
  radar_margins = FALSE,
  ...
)
}
\arguments{
\item{plotlist}{A list of plots}

\item{align}{Alignment character value}

\item{label_fontface}{Font face}

\item{label_size}{Label size}

\item{hjust}{Horizontal adjustment value}

\item{radar_margins}{Logical for margins that work well with \code{\link[=plot_radar]{plot_radar()}}}

\item{...}{Other arguments to pass to \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}. In particular,
you will probably want to use the \code{labels} argument.}
}
\value{
A ggplot object
}
\description{
This is a slightly customized wrapper for \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}.
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
g <- list()
g[[1]] <- plot_radar(probs)
g[[2]] <- plot_radar(probs)
plot_grid_pbs(g, labels = c("First plot", "Second plot"), radar_margins = TRUE)
}
