% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-index.R
\name{plot_index}
\alias{plot_index}
\title{Plot sampled survey index projections}
\usage{
plot_index(object_list, n_samples = 5, seed = 42, type = c("Ind", "AddInd"))
}
\arguments{
\item{object_list}{A list of DLMtool MSE objects representing different
scenarios. The list should be named with the scenario names.}

\item{n_samples}{The number of timeseries samples to illustrate.}

\item{seed}{The seed to set before drawing samples.}

\item{type}{Which index to plot.}
}
\value{
A ggplot object
}
\description{
Plot sampled survey index projections
}
\examples{
plot_index(mse_example)
mse_list <- list()
mse_list[[1]] <- mse_example
mse_list[[2]] <- mse_example
names(mse_list) <- c("Sc 1", "Sc 2")
plot_index(mse_list)
}
