% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-index-fit.R
\name{plot_index_fits}
\alias{plot_index_fits}
\title{Plot index fits from a stock reduction analysis}
\usage{
plot_index_fits(
  sra_list,
  survey_names,
  sample_n = min(c(100, sra_list[[1]]@OM@nsim)),
  alpha = 0.05,
  french = isTRUE(getOption("french"))
)
}
\arguments{
\item{sra_list}{A named a list of objects from \code{\link[SAMtool:RCM]{SAMtool::RCM()}}.}

\item{survey_names}{A character vector of real index names. Columns will be
ordered by the order of the survey names.}

\item{sample_n}{The number of SRA samples to plot.}

\item{alpha}{Transparency}

\item{french}{French?}
}
\value{
A ggplot2 object.
}
\description{
Plot index fits from a stock reduction analysis
}
\examples{
library(SAMtool)
om <- MSEtool::testOM
om@nsim <- 10
set.seed(1)
m <- RCM(om,
  data = list(Chist = runif(10), Index = runif(10), I_sd = rep(0.1, 10))
)
fits <- list()
fits[[1]] <- m
fits[[2]] <- m
names(fits) <- c("Sc 1", "Sc 2")
plot_index_fits(fits, survey_names = c("a"), alpha = 0.3)
}
