% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-kobe.R
\name{plot_kobe}
\alias{plot_kobe}
\title{Make a panel of contour plots by management procedure for B/BMSY by F/FMSY}
\usage{
plot_kobe(
  object,
  yend = max(object@proyears),
  dontshow_mp = NULL,
  show_ref_pt_lines = TRUE,
  alpha = c(0.25, 0.5, 0.75),
  n = 200,
  xlim = c(0, 3.5),
  ylim = c(0, 3.5),
  x_ref_lines = c(0.4, 0.8),
  y_ref_lines = 1,
  show_contours = TRUE,
  return_data = FALSE
)
}
\arguments{
\item{object}{An MSE object as returned by \code{\link[DLMtool:runMSE]{DLMtool::runMSE()}}.}

\item{yend}{The end year}

\item{dontshow_mp}{A vector of MPs to leave out of the plot.}

\item{show_ref_pt_lines}{Show the reference point lines at the values in *_ref_lines arguments}

\item{alpha}{A vector of levels between 0 and 1 for the contour lines.}

\item{n}{As defined in \code{\link[MASS:kde2d]{MASS::kde2d()}}.}

\item{xlim}{The \code{\link[ggplot2:xlim]{ggplot2::xlim()}} values for x-axis limits.}

\item{ylim}{The \code{\link[ggplot2:ylim]{ggplot2::ylim()}} values for y-axis limits.}

\item{x_ref_lines}{A vector of vertical lines to draw as reference point lines.}

\item{y_ref_lines}{A vector of horizontal lines to draw as reference point lines.}

\item{show_contours}{Logical: show contour lines?}

\item{return_data}{Logical: return the data instead of the plot?}
}
\value{
A ggplot object
}
\description{
Make a panel of contour plots by management procedure for B/BMSY by F/FMSY
}
\examples{
plot_kobe(mse_example)
}
