% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Make lollipop plot}
\usage{
plot_lollipop(pm_df_list, custom_pal = NULL, dodge = 0.6, pt_size = 2)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames from
\code{\link[=get_probs]{get_probs()}}. The names will be used as the plot labels.
Can also be a single data frame from \code{\link[=get_probs]{get_probs()}}.}

\item{custom_pal}{An optional custom color palette. Should be a named
character vector}

\item{dodge}{The amount to separate or "dodge" the lollipop lines.}

\item{pt_size}{Point size.}
}
\value{
A ggplot2 object
}
\description{
Make lollipop plot
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm <- list()
pm[[1]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm[[2]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
names(pm) <- c("Scenario 1", "Scenario 2")
plot_lollipop(pm)
plot_lollipop(pm[1]) + ggplot2::scale_colour_brewer(palette = "Set2")
}
