% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-parallel-coordinates.R
\name{plot_parallel_coords}
\alias{plot_parallel_coords}
\title{Make parallel coordinates plot of performance metrics}
\usage{
plot_parallel_coords(
  pm_df_list,
  type = c("facet", "single"),
  custom_pal = NULL,
  groups = NULL,
  rotate_labels = type == "facet",
  french = isTRUE(getOption("french"))
)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames from
\code{\link[=get_probs]{get_probs()}}. The names will be used as the plot labels.}

\item{type}{The type of plot. Multipanel \code{"facet"} vs. single panel
\code{"single"}. In the single panel version, a shaded ribbon represents the
upper and lower values across the scenarios and the line represents the
mean.}

\item{custom_pal}{An optional custom color palette. Should be a named
character vector}

\item{groups}{An optional grouping structure to separate the lines. For
example, this allows for conservation and fisheries performance metrics to
use distinct line segments. Should be a list with the performance metrics
grouped. See the example below.}

\item{rotate_labels}{Logical: rotate the performance metric labels 90
degrees?}

\item{french}{French?}
}
\value{
A ggplot2 object
}
\description{
Make parallel coordinates plot of performance metrics
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm <- list()
pm[[1]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm[[2]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
names(pm) <- c("Scenario 1", "Scenario 2")
plot_parallel_coords(pm)
plot_parallel_coords(pm, "single")
plot_parallel_coords(pm, groups = list(c("P40", "P100", "PNOF"), c("LTY", "AAVY")))
}
