% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-tables.R
\name{plot_probs}
\alias{plot_probs}
\title{Summary of probabilities of things from the MSE object in a colored tile table format}
\usage{
plot_probs(
  probs_dat,
  digits = 2,
  relative_max = FALSE,
  scale_0_1 = FALSE,
  mp_order = NULL,
  sort_by = "decreasing"
)
}
\arguments{
\item{probs_dat}{A list of length 2 - a data frame and another list of captions describing the
columns of the data frame as returned from \code{\link[gfdlm:get_probs]{gfdlm::get_probs()}}}

\item{digits}{How many decimal places to show in the tiles for the values}

\item{relative_max}{Make the plot have each column use a reletive maximum. If
scale_0_1 is used, this will be ignored}

\item{scale_0_1}{Scale each column from 0 to 1, so that the colours in each column are fully represented}

\item{mp_order}{Optional hardcoded MP order}

\item{sort_by}{show values in decreasing or increasing format}
}
\description{
Summary of probabilities of things from the MSE object in a colored tile table format
}
\examples{
library(ggplot2)
probs <- get_probs(mse, "P40", "P100", "PNOF", "LTY", "AAVY")
plot_probs(probs)
}
