% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-radar.R
\name{plot_radar_facet}
\alias{plot_radar_facet}
\title{Make a set of radar plots}
\usage{
plot_radar_facet(
  pm_df_list,
  custom_pal = NULL,
  ncol = NULL,
  nrow = NULL,
  label_size = 12,
  ...
)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames from \code{\link[=get_probs]{get_probs()}}. The names will be used as the plot labels.}

\item{custom_pal}{An optional custom color palette. Will be fed to \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}

\item{ncol}{An optional number of columns in the grid.}

\item{nrow}{An optional number of rows in the grid.}

\item{label_size}{Label size for the plots.}

\item{...}{Other arguments to pass to \code{\link[=plot_radar]{plot_radar()}}.}
}
\value{
A ggplot2 object
}
\description{
Make a set of radar plots
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm <- list()
pm[[1]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm[[2]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
names(pm) <- c("Scenario 1", "Scenario 2")
plot_radar_facet(pm)
}
