% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-rcm.R
\name{plot_rcm}
\alias{plot_rcm}
\alias{plot_rcm_SSB}
\alias{plot_rcm_SSBMSY}
\alias{plot_rcm_depletion}
\alias{plot_rcm_F}
\alias{plot_rcm_FMSY}
\alias{plot_rcm_recdev}
\alias{plot_rcm_rec}
\alias{plot_rcm_bio_sel}
\alias{plot_rcm_sel}
\alias{plot_rcm_sel_multi}
\alias{plot_rcm_corr}
\title{Plotting functions from OM conditioning}
\usage{
plot_rcm_SSB(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed",
  ylim = NULL,
  MPD = FALSE
)

plot_rcm_SSBMSY(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed",
  histogram = FALSE,
  bins = NULL,
  MPD = FALSE
)

plot_rcm_depletion(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed",
  histogram = FALSE,
  bins = NULL,
  MPD = FALSE
)

plot_rcm_F(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed",
  ylim = NULL,
  MPD = FALSE
)

plot_rcm_FMSY(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed",
  ylim = NULL,
  MPD = FALSE
)

plot_rcm_recdev(
  rcm,
  scenario,
  french = FALSE,
  proj = FALSE,
  logspace = FALSE,
  scales = "fixed",
  ylim = NULL,
  MPD = FALSE
)

plot_rcm_rec(
  rcm,
  scenario,
  french = FALSE,
  scales = "fixed",
  ylim = NULL,
  MPD = FALSE
)

plot_rcm_bio_sel(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  bio_type = c("LAA", "mat"),
  sel_type = c("fleet", "index"),
  sel_i = 1,
  sel_name = "Selectivity"
)

plot_rcm_sel(
  rcm,
  scenario,
  type = c("fleet", "index"),
  i = "all",
  color = "black",
  french = FALSE,
  MPD = FALSE
)

plot_rcm_sel_multi(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  fleet_i = 1,
  index_i = 1,
  fleet_names = paste("Fleet", fleet_i),
  index_names = paste("Index", index_i),
  french = FALSE,
  color
)

plot_rcm_corr(
  rcm,
  scenario,
  var = c("D", "M", "h", "R0"),
  status = TRUE,
  bins = NULL,
  do_corr = FALSE,
  french = FALSE
)
}
\arguments{
\item{rcm}{A list containing \linkS4class{RCModel} objects.}

\item{scenario}{A character vector of names corresponding to \code{rcm}.}

\item{french}{Logical, whether the axes are in French or not.}

\item{scales}{The scales argument to \link[ggplot2]{facet_wrap}.}

\item{ylim}{Optional y-axes limits to figures.}

\item{MPD}{Logical, whether to plot individual simulations (\code{FALSE}) or from the single fit (\code{TRUE}) in \code{RCModel@mean_fit}.}

\item{histogram}{Logical, whether to show a histogram (TRUE) for the terminal year only. Otherwise, a time series plot.}

\item{bins}{The number of bins for \link[ggplot]{geom_histogram}.}

\item{proj}{Logical, whether to plot the recruitment deviates in the projection period}

\item{logspace}{Logical, whether to plot the recruitment deviates in logspace or normal space}

\item{bio_type}{Character vector to plot either length-at-age (\code{"LAA"}) or maturity-at-age (\code{"mat"})}

\item{sel_type}{Character to indicate whether to grab a fleet or index selectivity}

\item{sel_i}{The index of the fleet or index for the selectivity}

\item{sel_name}{The name for the selectivity}

\item{type}{Whether the output is from a fishing fleet or index.}

\item{i}{The corresponding index from the fleet or index of abundance.}

\item{color}{Character vector of colors used for plotting.}

\item{fleet_i}{Vectors of indices for the fleets, i.e., 1 is the first fleet, etc. Use NA if no fleets will be plotted.}

\item{index_i}{Vectors of indices for the indices of abundance. Use NA if no indices will be plotted.}

\item{fleet_names}{Character vector for the names of the fleets}

\item{index_names}{Character vector for the names of the indices}

\item{var}{Parameters to plot. Select up to two.}

\item{status}{Logical, whether to identify estimated status with each point in scatter plot.
Only used if two parameters are selected in \code{var}.}

\item{do_corr}{Whether to report the correlation between parameters.
Only used if two parameters are selected in \code{var}.}
}
\description{
A set of functions that plots output from a list of RCModel objects used to condition a set
of operating models.
}
\section{Functions}{
\itemize{
\item \code{plot_rcm_SSB()}: Plot historical SSB

\item \code{plot_rcm_SSBMSY()}: Plot SSB/SSBMSY

\item \code{plot_rcm_depletion()}: Plot historical spawning depletion

\item \code{plot_rcm_F()}: Plot historical apical F

\item \code{plot_rcm_FMSY()}: Plot historical F/FMSY

\item \code{plot_rcm_recdev()}: Plot recruitment deviates

\item \code{plot_rcm_rec()}: Plot recruitment

\item \code{plot_rcm_bio_sel()}: Compare maturity and length-at-age schedule vs. selectivity (of a single index or fleet)

\item \code{plot_rcm_sel()}: Plot selectivity of a single index or fleet (all simulations)

\item \code{plot_rcm_sel_multi()}: Plot selectivity of a multiple indices or fleets overlayed on top of each other

\item \code{plot_rcm_corr()}: Plot histogram or scatter plot of operating model parameters

}}
