% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-rcm-data.R
\name{plot_rcm_data}
\alias{plot_rcm_data}
\alias{plot_rcm_index}
\alias{plot_rcm_age_comps}
\alias{plot_rcm_mean_age}
\alias{plot_rcm_length_comps}
\alias{plot_rcm_mean_length}
\alias{plot_rcm_biomass_age}
\title{Plotting functions of data used for OM conditioning}
\usage{
plot_rcm_index(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  i,
  index_names,
  color,
  MPD = FALSE
)

plot_rcm_age_comps(
  RCModel,
  scenario,
  french = FALSE,
  type = c("fleet", "index"),
  i = 1,
  color = "black",
  MPD = FALSE
)

plot_rcm_mean_age(
  rcm,
  scenario,
  type = c("fleet", "index"),
  i = 1,
  scales = "fixed",
  color = "black",
  alpha = 0.6,
  french = FALSE,
  MPD = FALSE
)

plot_rcm_length_comps(
  RCModel,
  scenario,
  french = FALSE,
  type = c("fleet", "index"),
  i = 1,
  color = "black",
  MPD = FALSE
)

plot_rcm_mean_length(
  rcm,
  scenario,
  type = c("fleet", "index"),
  i = 1,
  scales = "fixed",
  color = "black",
  alpha = 0.6,
  french = FALSE,
  MPD = FALSE
)

plot_rcm_biomass_age(
  RCModel,
  scenario,
  type = c("fleet", "index"),
  i = 1,
  color = "black",
  french = FALSE
)
}
\arguments{
\item{rcm}{A list containing \linkS4class{RCModel} objects.}

\item{scenario}{A character vector of names corresponding to \code{rcm}.}

\item{french}{Logical, whether the axes are in French or not.}

\item{i}{The corresponding index from the fleet or index of abundance.}

\item{index_names}{Character vector of names for the indices of abundance.}

\item{color}{Character vector of colors used for plotting.}

\item{MPD}{Logical, whether to plot individual simulations (\code{FALSE}) or from the single fit (\code{TRUE}) in
\code{RCModel@mean_fit}.}

\item{RCModel}{An object of class \linkS4class{RCModel}}

\item{type}{Whether the output is from a fishing fleet or index.}

\item{scales}{The scales argument to \link[ggplot2]{facet_wrap}.}

\item{alpha}{The alpha parameter to control color transparency in the ggplots.}
}
\description{
A set of functions that plots fits to data from a list of RCModel objects.
}
\section{Functions}{
\itemize{
\item \code{plot_rcm_index()}: Plot fits to index

\item \code{plot_rcm_age_comps()}: Plot age comps from a fleet or index (single RCM only)

\item \code{plot_rcm_mean_age()}: Plot mean age for a fleet or index of abundance

\item \code{plot_rcm_length_comps()}: Plot length comps from a fleet or index (single RCM only)

\item \code{plot_rcm_mean_length()}: Plot mean length for a fleet or index of abundance using the length compositions

\item \code{plot_rcm_biomass_age()}: Plots the proportion of biomass vs. age from the predicted age composition of a
fleet or index. Used as a diagnostic of whether a high residual in the plusgroup is problematic. Only plots
years for which there are observed age data.

}}
