% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-projections-scenarios2.R
\name{plot_scenario}
\alias{plot_scenario}
\alias{plot_scenario_BMSY}
\alias{plot_scenario_SSB}
\alias{plot_scenario_F}
\alias{plot_scenario_FMSY}
\alias{plot_scenario_catch}
\title{Plot a single time series for various scenarios and MPs}
\usage{
plot_scenario_BMSY(
  object_list,
  MP = NULL,
  quantiles = c(0.025, 0.975),
  facet = c("wrap", "grid"),
  palette = "Dark2",
  n_samples = 0,
  seed = 42,
  french = isTRUE(getOption("french"))
)

plot_scenario_SSB(
  object_list,
  MP = NULL,
  quantiles = c(0.025, 0.975),
  facet = c("wrap", "grid"),
  palette = "Dark2",
  n_samples = 2,
  seed = 42,
  french = isTRUE(getOption("french"))
)

plot_scenario_F(
  object_list,
  MP = NULL,
  quantiles = c(0.025, 0.975),
  facet = c("wrap", "grid"),
  palette = "Dark2",
  n_samples = 2,
  seed = 42,
  french = isTRUE(getOption("french"))
)

plot_scenario_FMSY(
  object_list,
  MP = NULL,
  quantiles = c(0.025, 0.975),
  facet = c("wrap", "grid"),
  palette = "Dark2",
  n_samples = 2,
  seed = 42,
  french = isTRUE(getOption("french"))
)

plot_scenario_catch(
  object_list,
  MP = NULL,
  quantiles = c(0.025, 0.975),
  facet = c("wrap", "grid"),
  palette = "Dark2",
  n_samples = 2,
  seed = 42,
  french = isTRUE(getOption("french"))
)
}
\arguments{
\item{object_list}{A named list of MSE objects from MSEtool. Names become scenario names.}

\item{MP}{Optional subset of management procedures}

\item{quantiles}{Length 2 vector for the quantiles to plot as ribbons}

\item{facet}{Character to indicate a wrapped facet plot by MP or grid by OM and MP. See description.}

\item{palette}{The name of an RColorBrewer colour palette}

\item{n_samples}{The number of timeseries samples to illustrate. Only used if \code{facet = "grid"}.}

\item{seed}{The seed to set before drawing samples}

\item{french}{Logical to use French translation.}
}
\description{
Plot historical and projected time series for various scenarios (B, B/BMSY, catch, etc.) either
(1) separate panels by MP that overlap by OM scenario, or (2) as a full grid by OM scenario and MP.
See also \code{\link[=plot_scenario_index]{plot_scenario_index()}}.
}
