% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-projections-scenarios.R
\name{plot_scenario_projections}
\alias{plot_scenario_projections}
\title{Plot projections for various scenarios on the same panels}
\usage{
plot_scenario_projections(
  object_list,
  probs = 0.5,
  rel_widths = c(2, 1.2),
  palette = "Dark2",
  catch_breaks = NULL,
  catch_labels = NULL,
  catch_ylim = NULL,
  msy_ylim = c(0, 4.5),
  french = isTRUE(getOption("french"))
)
}
\arguments{
\item{object_list}{A named list of MSE objects from DLMtool. Names become
scenario names.}

\item{probs}{The quantiles to plot as ribbons. This represents the tail
probability.}

\item{rel_widths}{A numeric vector of length 2 the controls the relative
width of the SSB and F columns (first element of the numeric vector) and
the catch column (second element of the numeric vector). Depending on the
number of years in the projection on the size of the axis labels, you may
need to tweak the second element of this vector to make all 3 columns the
same width. Figured it out by trial and error.}

\item{palette}{The name of an \pkg{RColorBrewer} colour palette.}

\item{catch_breaks}{Optional y-axis tick locations for the catch column.}

\item{catch_labels}{Optional y-axis tick labels for the catch column. Helpful
for dealing with large numbers.}

\item{catch_ylim}{An optional numeric factor of length two that represents
the y axis limits for the catch panels.}

\item{msy_ylim}{SSB and F column y limits.}

\item{french}{French?}
}
\value{
A ggplot object
}
\description{
Plot projections for various scenarios on the same panels
}
\examples{
x <- list()
x[[1]] <- mse_example
x[[2]] <- mse_example
names(x) <- c("Scenario 1", "Scenario 2")
plot_scenario_projections(x)
plot_scenario_projections(mse_example)
}
