% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-tigure.R
\name{plot_tigure_facet}
\alias{plot_tigure_facet}
\title{Make a set of tigure plots}
\usage{
plot_tigure_facet(pm_df_list, ncol = NULL, ...)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames from \code{\link[=get_probs]{get_probs()}}. The names will be used as the plot labels.}

\item{ncol}{Optional number of columns.}

\item{...}{Other arguments to pass to \code{\link[=plot_tigure]{plot_tigure()}}.}
}
\value{
A ggplot2 object
}
\description{
Make a set of tigure plots
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm <- list()
pm[[1]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm[[2]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
names(pm) <- c("Scenario 1", "Scenario 2")
plot_tigure_facet(pm)
}
