% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-bivariate.R
\name{plot_tradeoff}
\alias{plot_tradeoff}
\title{Make a trade-off scatterplot}
\usage{
plot_tradeoff(pm_df_list, xvar, yvar, custom_pal = NULL, mp = NULL)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames from
\code{\link[=get_probs]{get_probs()}}. The names will be used as the plot labels.}

\item{xvar}{The performance metric for the x axis (as character).}

\item{yvar}{The performance metric for the y axis (as character).}

\item{custom_pal}{An optional custom color palette. Should be a named
character victor.}

\item{mp}{An optional character vector of MPs to include. By default includes
all.}
}
\value{
A ggplot2 object
}
\description{
Creates a trade-off scatterplot of 2 performance metrics across multiple
scenarios
}
\examples{
probs <- list(get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY"))
names(probs) <- "Scenario 1"
plot_tradeoff(probs, "P40", "LTY")
}
