% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-worm.R
\name{plot_worms}
\alias{plot_worms}
\title{Make a Kobe "worm" timeseries plot with uncertainty}
\usage{
plot_worms(object, this_year, prob = 0.5, include_historical = TRUE)
}
\arguments{
\item{object}{An MSE object from DLMtool}

\item{this_year}{This year as a numeric value}

\item{prob}{Tail probability for the quantiles. 0.5 refers to an
interquartile range.}

\item{include_historical}{Logical: include the historical time?}
}
\value{
A ggplot object
}
\description{
Make a Kobe "worm" timeseries plot with uncertainty
}
\examples{
plot_worms(mse_example, this_year = 2020)
}
