% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{pm_factory}
\alias{pm_factory}
\title{Function factory for creating DLMtool performace metric (class "PM") functions}
\usage{
pm_factory(pm_type, ref = 0.1, yrs = NULL)
}
\arguments{
\item{pm_type}{The type of performance metric}

\item{ref}{Reference level to use in secondary part of probability}

\item{yrs}{A vector of years to include. If NULL, all will be used.}
}
\value{
A DLMtool PM function
}
\description{
Function factory for creating DLMtool performace metric (class "PM") functions
}
\examples{
P10 <- pm_factory("SBMSY", 0.1)
}
