% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{PNOF}
\alias{PNOF}
\title{The function used for Probaiblity of Not Overfishing Performance metric}
\usage{
PNOF(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for Probaiblity of Not Overfishing Performance metric
}
