% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{PNOF_yrs6_20}
\alias{PNOF_yrs6_20}
\title{The function used for Probaiblity of Not Overfishing Performance metric for years 6-20}
\usage{
PNOF_yrs6_20(mse_obj)
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
The function used for Probaiblity of Not Overfishing Performance metric for years 6-20
}
