% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-kobe.R
\name{quantile_contour}
\alias{quantile_contour}
\title{Calculate a set of x-y coordinates for a contour line
Density values are sorted and standardized, and the critical value is
calculated from the alpha cutoff value and sent to \code{\link[gfdlm:calc_contour_lines]{gfdlm::calc_contour_lines()}}
as an argument.}
\usage{
quantile_contour(x, alpha = 0.8)
}
\arguments{
\item{x}{Output from the function \code{\link[MASS:kde2d]{MASS::kde2d()}}}

\item{alpha}{A cutoff value between 0 and 1 for the contour line.}
}
\value{
A list of two vectors of x and y-coordinates for the contour line.
}
\description{
Calculate a set of x-y coordinates for a contour line
Density values are sorted and standardized, and the critical value is
calculated from the alpha cutoff value and sent to \code{\link[gfdlm:calc_contour_lines]{gfdlm::calc_contour_lines()}}
as an argument.
}
