% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{reduce_survey}
\alias{reduce_survey}
\title{Eliminate survey data for selected years}
\usage{
reduce_survey(mp, slots = c("Ind", "CAA", "CAL"), index = function(x)
  seq(1, x, by = 2))
}
\arguments{
\item{mp}{An existing management procedure function of class \code{"MP"} that will
work with \pkg{DLMtool}.}

\item{slots}{The slots for which you want to reduce observations.}

\item{index}{A function that takes the number of years of observations and
returns a vector indexing the years that should be turned into \code{NA} values.
The default anonymous function discards observations for odd years.}
}
\value{
A management procedure function of class \code{"MP"} for use with
\pkg{DLMtool}.
}
\description{
This function factory removes the selected years of observations before
feeding the data to an existing management procedure function in
\pkg{DLMtool}. For example, this can be used to turn an annual survey into a
biennial or triennial survey. It could also be used to eliminate commercial
observations from some years.
}
\examples{
library(DLMtool)
om <- DLMtool::testOM
om@nsim <- 3
temp_mp <- reduce_survey(Islope1)
mse <- runMSE(OM = om, MPs = "temp_mp")
}
