% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{SBT1_log}
\alias{SBT1_log}
\title{SBT simple MP in log space}
\usage{
SBT1_log(
  x,
  Data,
  reps = 100,
  plot = FALSE,
  yrsmth = 10,
  k1 = 1.5,
  k2 = 3,
  gamma = 1
)
}
\arguments{
\item{x}{A position in the data object. As per \pkg{DLMtool}.}

\item{Data}{A data object. As per \pkg{DLMtool}.}

\item{reps}{The number of stochastic samples of the MP recommendation(s). As
per \pkg{DLMtool}.}

\item{plot}{Logical.}

\item{yrsmth}{The number of years for evaluating trend in relative abundance indices.}

\item{k1}{Control parameter.}

\item{k2}{Control parameter.}

\item{gamma}{Control parameter.}
}
\description{
This MP is based on \code{\link[DLMtool:SBT1]{DLMtool::SBT1()}} but fits the linear regression in log space. SBT stands for "southern bluefin tuna".
}
