% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-spider.R
\name{spider}
\alias{spider}
\title{Create a spider plot}
\usage{
spider(
  mse_obj,
  pm_list = list("LTY", "P10", "P40", "P80", "P100", "PNOF", "AAVY"),
  palette = "Set2",
  ...
)
}
\arguments{
\item{mse_obj}{A DLMtool object of class \linkS4class{MSE}}

\item{pm_list}{A list of strings defining valid Performance metrics}

\item{palette}{A palette color as recognized by \code{\link[ggplot2:scale_color_brewer]{ggplot2::scale_color_brewer()}}}

\item{...}{Other arguments to pass to \code{\link[ggspider:spider_web]{ggspider::spider_web()}}.}
}
\value{
A ggplot object
}
\description{
Create a spider plot
}
\examples{
spider(mse)
}
