% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_pbs}
\alias{theme_pbs}
\title{A ggplot2 theme}
\usage{
theme_pbs(
  base_size = 11,
  base_family = "",
  text_col = "grey20",
  panel_border_col = "grey70"
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}

\item{text_col}{Color for text.}

\item{panel_border_col}{Color for panel borders.}
}
\description{
A simple theme for ggplot2 that loosely resembles nicely themed plots from
base graphics.
}
\examples{
p <- ggplot2::ggplot(mtcars) +
  ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = factor(gear))) +
  ggplot2::facet_wrap(~am)
p + theme_pbs()
}
