% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_cal}
\alias{tidy_cal}
\alias{tidy_caa}
\title{Generate length-at-age or catch-at-age data}
\usage{
tidy_cal(dat, yrs, unsorted_only = TRUE, interval = 1, sex = c(1, 2))

tidy_caa(dat, yrs, unsorted_only = FALSE, interval = 1, sex = c(1, 2))
}
\arguments{
\item{dat}{Commercial or biological samples from \code{\link[=get_commercial_samples]{get_commercial_samples()}} or
\code{\link[=get_survey_samples]{get_survey_samples()}}.}

\item{yrs}{A complete set of years to include in the matrix.}

\item{unsorted_only}{Logical for whether to only include the unsorted samples.
Only applies to the commercial data.}

\item{interval}{Interval for the complete set of ages or lengths. For example,
for length bins of interval 2, \code{interval = 2}.}

\item{sex}{Numeric vector indicating males and/or or females.}
}
\value{
A catch at age or catch at length matrix as an array.
1 x nyears x nage/nlength
}
\description{
Generate length-at-age or catch-at-age data
}
\examples{
\donttest{
d_commercial <- gfdata::get_commercial_samples(222)
cal <- tidy_cal(d_commercial, yrs = 2005:2010, interval = 5)

d_survey <- gfdata::get_survey_samples(222, ssid = 1)
caa <- tidy_caa(d_survey, yrs = 2010:2012)
caa[1, , ]

}
}
