% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_mean_length}
\alias{tidy_mean_length}
\title{Generate mean-length time series}
\usage{
tidy_mean_length(dat, unsorted_only = FALSE)
}
\arguments{
\item{dat}{Commercial or biological samples from \code{\link[=get_commercial_samples]{get_commercial_samples()}} or
\code{\link[=get_survey_samples]{get_survey_samples()}}.}

\item{unsorted_only}{Logical for whether to only include the unsorted samples.
Only applies to the commercial data.}
}
\description{
Generate mean-length time series
}
