% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-tables.R
\name{get_probs}
\alias{get_probs}
\title{Summary of probabilities of performance metrics from the MSE object}
\usage{
get_probs(object, ..., refs = NULL)
}
\arguments{
\item{object}{MSE object, output of the \code{\link[DLMtool:runMSE]{DLMtool::runMSE()}} function}

\item{...}{List of performace metrics}

\item{refs}{List containing the reference limits for each metric}
}
\value{
A tibble of the output
}
\description{
Summary of probabilities of performance metrics from the MSE object
}
\examples{
library(DLMtool)
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
}
