% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{CC_hist20}
\alias{CC_hist20}
\alias{GB_slope6_0.66}
\alias{.GB_slope6_0.66}
\alias{GB_slope6_1}
\alias{.GB_slope6_1}
\alias{GB_slope8_0.66}
\alias{.GB_slope8_0.66}
\alias{GB_slope8_1}
\alias{.GB_slope8_1}
\alias{.ICI}
\alias{.ICI2}
\alias{Iratio2}
\alias{.Iratio2}
\alias{Islope0.4_100}
\alias{Islope0.4_80}
\alias{Islope0.2_100}
\alias{Islope0.2_80}
\alias{IT5_hist}
\alias{IT10_hist}
\alias{CC1.0}
\alias{CC0.9}
\alias{CC0.8}
\alias{CC0.7}
\alias{CC0.6}
\alias{CC1.1}
\alias{CC1.2}
\title{Alternative versions of \pkg{DLMtool} and \pkg{MSEtool} MP functions}
\usage{
CC_hist20(x, Data, reps = 100, ...)

GB_slope6_0.66(x, Data, reps = 100, ...)

.GB_slope6_0.66(x, Data, reps = 100, ...)

GB_slope6_1(x, Data, reps = 100, ...)

.GB_slope6_1(x, Data, reps = 100, ...)

GB_slope8_0.66(x, Data, reps = 100, ...)

.GB_slope8_0.66(x, Data, reps = 100, ...)

GB_slope8_1(x, Data, reps = 100, ...)

.GB_slope8_1(x, Data, reps = 100, ...)

.ICI(x, Data, reps = 100, ...)

.ICI2(x, Data, reps = 100, ...)

Iratio2(x, Data, reps = 100, ...)

.Iratio2(x, Data, reps = 100, ...)

Islope0.4_100(x, Data, reps = 100, ...)

Islope0.4_80(x, Data, reps = 100, ...)

Islope0.2_100(x, Data, reps = 100, ...)

Islope0.2_80(x, Data, reps = 100, ...)

IT5_hist(x, Data, reps = 100, ...)

IT10_hist(x, Data, reps = 100, ...)

CC1.0(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = 0)

CC0.9(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = 0.1)

CC0.8(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = 0.2)

CC0.7(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = 0.3)

CC0.6(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = 0.4)

CC1.1(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = -0.1)

CC1.2(x, Data, reps = 100, plot = FALSE, yrsmth = 5, xx = -0.1)
}
\arguments{
\item{x}{A position in the data object. As per \pkg{MSEtool}.}

\item{Data}{A data object. As per \pkg{MSEtool}.}

\item{reps}{The number of stochastic samples of the MP recommendation(s). As
per \pkg{MSEtool}.}

\item{...}{Other arguments to pass to the MP function.}

\item{plot}{Logical. Show the plot?}

\item{yrsmth}{Years over which to calculate mean catches.}

\item{xx}{Parameter controlling the TAC. Mean catches are multiplied by \code{(1 - xx)}.}
}
\description{
Management procedures (MPs) with a \code{.} in front of their name use
\code{\link[=reduce_survey]{reduce_survey()}} to eliminate odd years of survey observations to reflect
the biennial nature of most groundfish surveys in British Columbia.
}
