% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{AvC20}
\alias{AvC20}
\alias{.DDSS_MSY}
\alias{.DDSS_4010}
\alias{GB_slope6_0.66}
\alias{.GB_slope6_0.66}
\alias{GB_slope6_1}
\alias{.GB_slope6_1}
\alias{GB_slope8_0.66}
\alias{.GB_slope8_0.66}
\alias{GB_slope8_1}
\alias{.GB_slope8_1}
\alias{.ICI}
\alias{.ICI2}
\alias{Iratio6}
\alias{.Iratio6}
\alias{Iratio8}
\alias{.Iratio8}
\alias{Islope_mod1}
\alias{.Islope_mod1}
\alias{Islope_mod2}
\alias{Islope_mod3}
\alias{.Islope_mod3}
\alias{Islope_mod4}
\alias{.Islope_mod4}
\alias{.Islope1}
\alias{.Islope2}
\alias{.Islope3}
\alias{.Islope4}
\alias{.IT5}
\alias{.IT10}
\alias{.Itarget1}
\alias{.Itarget2}
\alias{.Itarget3}
\alias{.Itarget4}
\alias{.ITM}
\alias{.SBT1}
\alias{.IDX}
\alias{IDX_smooth}
\alias{.IDX_smooth}
\alias{SP4010}
\alias{.SP4010}
\alias{SP8040}
\alias{.SP8040}
\alias{SP8040_Fox}
\alias{.SP8040_Fox}
\alias{SP6040}
\alias{.SP6040}
\alias{SP6040_Fox}
\alias{.SP6040_Fox}
\alias{.SP_MSY}
\alias{SP_MSY_Fox}
\alias{.SP_MSY_Fox}
\alias{IT5_hist}
\alias{.IT5_hist}
\alias{IT10_hist}
\alias{.IT10_hist}
\title{Alternative versions of \pkg{DLMtool} and \pkg{MSEtool} MP functions}
\usage{
AvC20(x, Data, reps = 100, ...)

.DDSS_MSY(x, Data, reps = 100, ...)

.DDSS_4010(x, Data, reps = 100, ...)

GB_slope6_0.66(x, Data, reps = 100, ...)

.GB_slope6_0.66(x, Data, reps = 100, ...)

GB_slope6_1(x, Data, reps = 100, ...)

.GB_slope6_1(x, Data, reps = 100, ...)

GB_slope8_0.66(x, Data, reps = 100, ...)

.GB_slope8_0.66(x, Data, reps = 100, ...)

GB_slope8_1(x, Data, reps = 100, ...)

.GB_slope8_1(x, Data, reps = 100, ...)

.ICI(x, Data, reps = 100, ...)

.ICI2(x, Data, reps = 100, ...)

Iratio6(x, Data, reps = 100, ...)

.Iratio6(x, Data, reps = 100, ...)

Iratio8(x, Data, reps = 100, ...)

.Iratio8(x, Data, reps = 100, ...)

Islope_mod1(x, Data, reps = 100, ...)

.Islope_mod1(x, Data, reps = 100, ...)

Islope_mod2(x, Data, reps = 100, ...)

Islope_mod3(x, Data, reps = 100, ...)

.Islope_mod3(x, Data, reps = 100, ...)

Islope_mod4(x, Data, reps = 100, ...)

.Islope_mod4(x, Data, reps = 100, ...)

.Islope1(x, Data, reps = 100, ...)

.Islope2(x, Data, reps = 100, ...)

.Islope3(x, Data, reps = 100, ...)

.Islope4(x, Data, reps = 100, ...)

.IT5(x, Data, reps = 100, ...)

.IT10(x, Data, reps = 100, ...)

.Itarget1(x, Data, reps = 100, ...)

.Itarget2(x, Data, reps = 100, ...)

.Itarget3(x, Data, reps = 100, ...)

.Itarget4(x, Data, reps = 100, ...)

.ITM(x, Data, reps = 100, ...)

.SBT1(x, Data, reps = 100, ...)

.IDX(x, Data, reps = 100, ...)

IDX_smooth(x, Data, reps = 100, ...)

.IDX_smooth(x, Data, reps = 100, ...)

SP4010(x, Data, reps = 1)

.SP4010(x, Data, reps = 100, ...)

SP8040(x, Data, reps = 1)

.SP8040(x, Data, reps = 100, ...)

SP8040_Fox(x, Data, reps = 1)

.SP8040_Fox(x, Data, reps = 100, ...)

SP6040(x, Data, reps = 1)

.SP6040(x, Data, reps = 100, ...)

SP6040_Fox(x, Data, reps = 1)

.SP6040_Fox(x, Data, reps = 100, ...)

.SP_MSY(x, Data, reps = 100, ...)

SP_MSY_Fox(x, Data, reps = 1)

.SP_MSY_Fox(x, Data, reps = 100, ...)

IT5_hist(x, Data, reps = 100, ...)

.IT5_hist(x, Data, reps = 100, ...)

IT10_hist(x, Data, reps = 100, ...)

.IT10_hist(x, Data, reps = 100, ...)
}
\arguments{
\item{x}{A position in the data object. As per \pkg{DLMtool}.}

\item{Data}{A data object. As per \pkg{DLMtool}.}

\item{reps}{The number of stochastic samples of the MP recommendation(s). As
per \pkg{DLMtool}.}

\item{...}{Other arguments to pass to the MP function.}
}
\description{
Management procedures (MPs) with a \code{.} in front of their name use
\code{\link[=reduce_survey]{reduce_survey()}} to eliminate odd years of survey observations to reflect
the biennial nature of most groundfish surveys in British Columbia.
}
