% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-convergence.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot convergence}
\usage{
plot_convergence(object, pm_list = c("LTY", "PNOF"))
}
\arguments{
\item{object}{A DLMtool MSE object.}

\item{pm_list}{A list or vector of performance metrics as character.}
}
\value{
A ggplot2 plot.
}
\description{
This function plots cumulative performance metric values across iterations to
assess convergence.
}
\examples{
plot_convergence(mse)
}
