% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-convergence.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Plot convergence}
\usage{
plot_convergence(
  object,
  pm_list = c("LTY", "PNOF"),
  label_gap = 1.15,
  use_labels = FALSE,
  ylim = c(min(df$value), 1)
)
}
\arguments{
\item{object}{A DLMtool MSE object.}

\item{pm_list}{A list or vector of performance metrics as character.}

\item{label_gap}{Controls the gap on the right of each panel to make space
for the MP labels. This value is multiplied by the maximum number of
iterations to expand the x axis.}

\item{use_labels}{Use labels (\code{TRUE}) or use a colour legend (\code{FALSE})?}

\item{ylim}{Y limits. Defaults to the minimum observed performance metric
value (via lazy evaluation) and 1.}
}
\value{
A ggplot2 plot.
}
\description{
This function plots cumulative performance metric values across iterations to
assess convergence.
}
\examples{
plot_convergence(mse_example)
}
