% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-historical.R, R/figures-projections.R
\name{plot_historical_ts}
\alias{plot_historical_ts}
\title{Plot historical time series}
\usage{
plot_historical_ts(object, type = c("Catch", "SSB", "B", "VB",
  "Removals", "Rec", "N", "Find", "Marray", "RecDev"), n_samples = 50,
  this_year = 2018, observed_ts = NULL, scale = if
  (is.null(observed_ts)) FALSE else TRUE, legend_position = c(0.9, 0.85))

plot_historical_ts(object, type = c("Catch", "SSB", "B", "VB",
  "Removals", "Rec", "N", "Find", "Marray", "RecDev"), n_samples = 50,
  this_year = 2018, observed_ts = NULL, scale = if
  (is.null(observed_ts)) FALSE else TRUE, legend_position = c(0.9, 0.85))
}
\arguments{
\item{object}{A DLMtool object of class \code{Hist} that was created by
running \code{\link[DLMtool:runMSE]{DLMtool::runMSE()}} with \code{Hist = TRUE}.}

\item{type}{A character object describing the element of \code{object@TSdata} to
plot.}

\item{n_samples}{The number of timeseries samples to illustrate.}

\item{this_year}{The last year of the historical times there is.}

\item{observed_ts}{An optional observed timeseries to add is a comparison.}

\item{scale}{Should the timeseries be scaled by their geometric mean?}

\item{legend_position}{The x and y coordinates of the legend.}

\item{object}{A DLMtool object of class \code{mse}.}

\item{type}{A character object describing the element of \code{object@TSdata} to
plot.}

\item{n_samples}{The number of timeseries samples to illustrate.}

\item{this_year}{The last year of the historical times there is.}

\item{observed_ts}{An optional observed timeseries to add is a comparison.}

\item{legend_position}{The x and y coordinates of the legend. Only
gets used if an observed timeseries is also plotted.}
}
\value{
ggplot object

ggplot object
}
\description{
Plot historical time series

Plot projection timeseries
}
\examples{
library(DLMtool)
historical_mse <- runMSE(om, Hist = TRUE)
plot_historical_ts(historical_mse, n_samples = 2)
plot_historical_ts(historical_mse, n_samples = 2,
  observed_ts = rlnorm(50, 1, 0.3))

library(DLMtool)
historical_mse <- runMSE(om, Hist = TRUE)
plot_historical_ts(historical_mse, n_samples = 2)
plot_historical_ts(historical_mse, type = "SSB", n_samples = 2)
plot_historical_ts(historical_mse, type = "RecDev", n_samples = 2)
plot_historical_ts(historical_mse,
  n_samples = 2,
  observed_ts = rlnorm(50, 1, 0.3)
)
}
