% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-historical.R
\name{plot_historical_ts}
\alias{plot_historical_ts}
\title{Plot historical time series}
\usage{
plot_historical_ts(object, type = c("Catch", "SSB"), n_samples = 50,
  this_year = 2018, observed_ts = NULL, legend_position = c(0.9,
  0.85))
}
\arguments{
\item{object}{object}

\item{type}{type}

\item{n_samples}{number of samples}

\item{this_year}{the year}

\item{observed_ts}{obeserved ts}

\item{legend_position}{legend position}
}
\value{
ggplot object
}
\description{
Plot historical time series
}
