% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-historical.R
\name{plot_historical_ts}
\alias{plot_historical_ts}
\title{Plot historical time series}
\usage{
plot_historical_ts(object, type = c("Catch", "SSB"), n_samples = 50,
  this_year = 2018, observed_ts = NULL, legend_position = c(0.9,
  0.85))
}
\arguments{
\item{object}{A DLMtool object of class \code{MSE}.}

\item{type}{A character object describing the element of \code{object@TSdata} to
plot.}

\item{n_samples}{The number of timeseries samples to illustrate.}

\item{this_year}{The last year of the historical times there is.}

\item{observed_ts}{An optional observed timeseries to add is a comparison.}

\item{legend_position}{The x and y coordinates of the legend.}
}
\value{
ggplot object
}
\description{
Plot historical time series
}
