% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-index.R
\name{plot_index}
\alias{plot_index}
\alias{plot_scenario_index}
\alias{plot_catch}
\title{Plot sampled survey index or catch projections}
\usage{
plot_index(
  object_list,
  n_samples = 4,
  seed = 42,
  type = c("Ind", "AddInd"),
  MP = NULL,
  omit_index_fn = function(x) NULL,
  quantiles = c(0.025, 0.975),
  french = isTRUE(getOption("french"))
)

plot_scenario_index(
  object_list,
  seed = 42,
  type = c("Ind", "AddInd"),
  MP = NULL,
  omit_index_fn = function(x) NULL,
  quantiles = c(0.025, 0.975),
  palette = "Dark2",
  french = isTRUE(getOption("french"))
)

plot_catch(
  object_list,
  n_samples = 4,
  seed = 42,
  MP = NULL,
  quantiles = c(0.025, 0.975),
  french = isTRUE(getOption("french"))
)
}
\arguments{
\item{object_list}{A list of MSEtool MSE objects representing different
scenarios. The list should be named with the scenario names.}

\item{n_samples}{The number of timeseries samples to illustrate.}

\item{seed}{The seed to set before drawing samples.}

\item{type}{Which index to plot.}

\item{MP}{A character vector to subset the management procedures in the MSE object.}

\item{omit_index_fn}{A function that indexes years in the projection period
to omit from the plot. E.g., for a biennial survey omitting even years
starting in year 2: \verb{function(x) seq(2, x, 2))}. \code{x} represents the total
number of years.}

\item{quantiles}{Probability quantiles to show.}

\item{french}{French?}

\item{palette}{A palette color as recognized by \code{\link[ggplot2:scale_brewer]{ggplot2::scale_color_brewer()}}}
}
\value{
A ggplot object
}
\description{
\itemize{
\item \code{plot_index()}: Plots observed and simulated indices, i.e., \code{MSE@PPD[[i]]@Ind} or \code{MSE@PPD[[i]]@AddInd}, gridded by
OM and MP
}

\itemize{
\item \code{plot_scenario_index()}: Plots observed and simulated indices, i.e., \code{MSE@PPD[[i]]@Ind} or \code{MSE@PPD[[i]]@AddInd} by
MP (multiple OMs per panel)
}

\itemize{
\item \code{plot_catch()}:  Plots observed catches, i.e., \code{MSE@PPD[[i]]@Cat}. \strong{To plot OM catches, use \code{\link[=plot_scenario_catch]{plot_scenario_catch()}}}.
}
}
\examples{
plot_index(mse_example)
mse_list <- list()
mse_list[[1]] <- mse_example
mse_list[[2]] <- mse_example
names(mse_list) <- c("Sc 1", "Sc 2")
plot_index(mse_list)
}
