% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-parallel-coordinates.R
\name{plot_parallel_coords}
\alias{plot_parallel_coords}
\title{Make parallel coordinates plot of performance metrics}
\usage{
plot_parallel_coords(
  pm_df_list,
  type = c("facet", "single"),
  custom_pal = NULL,
  mp = NULL,
  rotate_labels = type == "facet"
)
}
\arguments{
\item{pm_df_list}{A named list of performance metric data frames from
\code{\link[=get_probs]{get_probs()}}. The names will be used as the plot labels.}

\item{type}{The type of plot. Multipanel \code{"facet"} vs. single panel
\code{"single"}. In the single panel version, a shaded ribbon represents the
upper and lower values across the scenarios.}

\item{custom_pal}{An optional custom color palette. Should be a named
character vector}

\item{mp}{An optional character vector of management procedures to filter to
include.}

\item{rotate_labels}{Logical: rotate the performance metric labels 90
degrees?}
}
\value{
A ggplot2 object
}
\description{
Make parallel coordinates plot of performance metrics
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm <- list()
pm[[1]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
pm[[2]] <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
names(pm) <- c("Scenario 1", "Scenario 2")
plot_parallel_coords(pm)
plot_parallel_coords(pm, "single")
}
