% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-projections.R
\name{plot_projection_ts}
\alias{plot_projection_ts}
\title{Plot projection time series}
\usage{
plot_projection_ts(
  object,
  type = c("SSB", "FM"),
  n_samples = 3,
  this_year = 2018,
  probs = c(0.1, 0.5),
  ribbon_colours = RColorBrewer::brewer.pal(8, "Blues")[c(2, 4, 8)],
  bbmsy_zones = c(0.4, 0.8),
  catch_reference = 1,
  clip_ylim = NULL,
  seed = 42
)
}
\arguments{
\item{object}{A DLMtool object of class \code{mse} that was created by
running \code{\link[DLMtool:runMSE]{DLMtool::runMSE()}}.}

\item{type}{A character vector describing the element of \verb{mse@} to
plot. Each of these "types" will be included as a column in the
final plot.}

\item{n_samples}{The number of timeseries samples to illustrate.}

\item{this_year}{The last year of the historical timeseries.}

\item{probs}{A numeric vector of the quantiles to plot as ribbons. Must be of length 2.}

\item{ribbon_colours}{A character vector of length three giving the colours for
the outer uncertainty ribbon, inner uncertainty ribbon, and median.}

\item{bbmsy_zones}{A numeric vector of status zone lines to add to
the B/Bmsy panel if it is present.}

\item{catch_reference}{The number of years over which to calculate the
average catch and show as a horizontal line in a plot of catch.}

\item{clip_ylim}{An optional numeric value to set the upper ylim
value. \code{clip_ylim} is multiplied by the highest median value
across the panels. Useful if there are outlying timeseries that distort
the y axis limits.}

\item{seed}{The seed to set before drawing samples.}
}
\value{
ggplot object
}
\description{
Plot projection time series
}
\examples{
plot_projection_ts(mse_example)
plot_projection_ts(mse_example, type = "SSB")
}
