% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-projections.R
\name{plot_projection_ts}
\alias{plot_projection_ts}
\title{Plot projection time series}
\usage{
plot_projection_ts(object, type = c("F_FMSY", "B_BMSY"), n_samples = 3,
  this_year = 2018, probs = c(0.1, 0.5),
  ribbon_colours = RColorBrewer::brewer.pal(8, "Blues")[c(2, 4, 8)],
  bbmsy_zones = c(0.4, 0.8))
}
\arguments{
\item{object}{A DLMtool object of class \code{mse} that was created by
running \code{\link[DLMtool:runMSE]{DLMtool::runMSE()}}.}

\item{type}{A character vector describing the element of \code{mse@} to
plot. Each of these "types" will be included as a column in the
final plot.}

\item{n_samples}{The number of timeseries samples to illustrate.}

\item{this_year}{The last year of the historical timeseries.}

\item{probs}{A numeric vector of the quantiles to plot as ribbons. Must be of length 2.}

\item{ribbon_colours}{A character vector of length three giving the colours for
the outer uncertainty ribbon, inner uncertainty ribbon, and median.}

\item{bbmsy_zones}{A numeric vector of status zone lines to add to
the B/Bmsy panel if it is present.}
}
\value{
ggplot object
}
\description{
Plot projection time series
}
\examples{
library(DLMtool)
om@nsim <- 10
x <- runMSE(om, MPs = c("AvC", "CC1"))
set.seed(1)
plot_projection_ts(x)
plot_projection_ts(x, type = "SSB")
}
