% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Create a radar plot}
\usage{
plot_radar(pm_df, palette = "Set2", ...)
}
\arguments{
\item{pm_df}{A performance metric data frame from \code{\link[=get_probs]{get_probs()}}.}

\item{palette}{A palette color as recognized by \code{\link[ggplot2:scale_color_brewer]{ggplot2::scale_color_brewer()}}}

\item{...}{Other arguments to pass to \code{\link[ggspider:spider_web]{ggspider::spider_web()}}.}
}
\value{
A ggplot object
}
\description{
Create a radar plot
}
\examples{
probs <- get_probs(mse_example, "P40", "P100", "PNOF", "LTY", "AAVY")
plot_radar(probs)
}
