% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-rcm.R
\name{plot_rcm}
\alias{plot_rcm}
\alias{plot_rcm_SSB}
\alias{plot_rcm_depletion}
\alias{plot_rcm_F}
\alias{plot_rcm_recdev}
\alias{plot_rcm_bio_sel}
\title{Plotting functions from OM conditioning}
\usage{
plot_rcm_SSB(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed",
  ylim = NULL
)

plot_rcm_depletion(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed"
)

plot_rcm_F(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  scales = "fixed",
  ylim = NULL
)

plot_rcm_recdev(
  rcm,
  scenario,
  french = FALSE,
  proj = FALSE,
  logspace = FALSE,
  scales = "fixed",
  ylim = NULL
)

plot_rcm_bio_sel(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  french = FALSE,
  bio_type = c("LAA", "mat"),
  sel_type = c("fleet", "index"),
  sel_i = 1,
  sel_name
)
}
\arguments{
\item{rcm}{A list containing \linkS4class{RCModel} objects.}

\item{scenario}{A character vector of names corresponding to \code{rcm}.}

\item{french}{Logical, whether the axes are in French or not.}

\item{scales}{The scales argument to \link[ggplot2]{facet_wrap}.}

\item{ylim}{Optional y-axes limits to figures.}

\item{proj}{Logical, whether to plot the recruitment deviates in the projection period}

\item{logspace}{Logical, whether to plot the recruitment deviates in logspace or normal space}

\item{bio_type}{Character vector to plot either length-at-age (\code{"LAA"}) or maturity-at-age (\code{"mat"})}

\item{sel_type}{Character to indicate whether to grab a fleet or index selectivity}

\item{sel_i}{The index of the fleet or index for the selectivity}

\item{sel_name}{The name for the selectivity}
}
\description{
A set of functions that plots output from a list of RCModel objects used to condition a set
of operating models.
}
\section{Functions}{
\itemize{
\item \code{plot_rcm_SSB}: Plot historical SSB

\item \code{plot_rcm_depletion}: Plot historical spawning depletion

\item \code{plot_rcm_F}: Plot historical apical F

\item \code{plot_rcm_recdev}: Plot recruitment deviates

\item \code{plot_rcm_bio_sel}: Compare maturity and length-at-age schedule vs. selectivity (of a single index or fleet)
}}

