% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-rcm-data.R
\name{plot_rcm_data}
\alias{plot_rcm_data}
\alias{plot_rcm_index}
\alias{plot_rcm_age_comps}
\alias{plot_rcm_sel}
\alias{plot_rcm_sel_multi}
\alias{plot_rcm_mean_age}
\alias{plot_rcm_length_comps}
\alias{plot_rcm_mean_length}
\alias{plot_rcm_biomass_age}
\title{Plotting functions of data used for OM conditioning}
\usage{
plot_rcm_index(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  FRENCH = FALSE,
  i,
  index_names,
  color,
  xlim
)

plot_rcm_age_comps(
  RCModel,
  scenario,
  FRENCH = FALSE,
  type = c("fleet", "index"),
  i = 1,
  color = "black"
)

plot_rcm_sel(rcm, scenario, type = c("survey", "index"), i, name)

plot_rcm_sel_multi(
  rcm,
  scenario = paste("Scenario", 1:length(rcm)),
  fleet_i = 1,
  index_i = 1,
  fleet_names = paste("Fleet", fleet_i),
  index_names = paste("Index", index_i),
  color
)

plot_rcm_mean_age(
  rcm,
  scenario,
  type = c("fleet", "index"),
  i,
  name,
  scales = "fixed",
  ylim
)

plot_rcm_length_comps(
  RCModel,
  scenario,
  FRENCH = FALSE,
  type = c("fleet", "index"),
  i = 1,
  color = "black"
)

plot_rcm_mean_length(
  rcm,
  scenario,
  type = c("fleet", "index"),
  i,
  name,
  scales = "fixed",
  ylim
)

plot_rcm_biomass_age(
  RCModel,
  scenario,
  type = c("fleet", "index"),
  i = 1,
  color = "black"
)
}
\arguments{
\item{rcm}{A list containing \linkS4class{RCModel} objects.}

\item{scenario}{A character vector of names corresponding to \code{rcm}.}

\item{FRENCH}{Logical, whether the axes are in French or not.}

\item{i}{The corresponding index from the fleet or index of abundance.}

\item{index_names}{Character vector for the names of the indices}

\item{color}{Character vector of colors used for plotting.}

\item{xlim}{Optional, x-axis limits for the figure.}

\item{RCModel}{An object of class \linkS4class{RCModel}}

\item{type}{Whether the output is from a fishing fleet or index.}

\item{name}{The name of the index or fleet}

\item{fleet_i}{Vectors of indices for the fleets, i.e., 1 is the first fleet, etc. Use NA if no fleets will be plotted.}

\item{index_i}{Vectors of indices for the indices of abundance. Use NA if no indices will be plotted.}

\item{fleet_names}{Character vector for the names of the fleets}

\item{scales}{The scales argument to \link[ggplot2]{facet_wrap}.}

\item{ylim}{Optional y-axes limits to figures.}
}
\description{
A set of functions that plots fits to data from a list of RCModel objects.
}
\section{Functions}{
\itemize{
\item \code{plot_rcm_index}: Plot fits to index

\item \code{plot_rcm_age_comps}: Plot age comps from a fleet or index (single RCM only)

\item \code{plot_rcm_sel}: Plot selectivity of a single index or fleet (all simulations)

\item \code{plot_rcm_sel_multi}: Plot selectivity of a multiple indices or fleets overlayed on top of each other

\item \code{plot_rcm_mean_age}: Plot mean age for a fleet or index of abundance

\item \code{plot_rcm_length_comps}: Plot length comps from a fleet or index (single RCM only)

\item \code{plot_rcm_mean_length}: Plot mean length for a fleet or index of abundance using the length compositions

\item \code{plot_rcm_biomass_age}: Plots the proportion of biomass vs. age from the predicted age composition of a
fleet or index. Used as a diagnostic of whether a high residual in the plusgroup is problematic. Only plots
years for which there are observed age data.
}}

