% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-sensitivity-trajectory.R
\name{plot_sensitivity_trajectory}
\alias{plot_sensitivity_trajectory}
\title{Sensitivity trajectory plot}
\usage{
plot_sensitivity_trajectory(
  object,
  type = c("B_BMSY", "F_FMSY"),
  mp = object@MPs,
  slots = c("D", "hs", "M", "ageM", "L50", "Linf", "K", "Isd")
)
}
\arguments{
\item{object}{A DLMtool MSE object.}

\item{type}{Whether to make a B/Bmsy or F/Fmsy projection plot.}

\item{mp}{A character vector of MPs to plot (must be included in the MSE
object). Defaults to all.}

\item{slots}{A character vector of OM slots to plot. Will be plotted in this
order. Set \code{slots = "all"} to plot all available OM and observation slots.}
}
\value{
A ggplot object
}
\description{
Plot sensitivity of performance metrics to OM slots across iterations
as trajectories.
}
\examples{
library(DLMtool)
om@nsim <- 10
x <- runMSE(om, MPs = c("AvC", "CC1"))
plot_sensitivity_trajectory(x)
}
