% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures-worm.R
\name{plot_worms}
\alias{plot_worms}
\title{plot_worms}
\usage{
plot_worms(object, prob = 0.5)
}
\arguments{
\item{object}{TODO}

\item{prob}{TODO}

\item{this_year}{TODO}
}
\value{
A ggplot object
}
\description{
plot_worms
}
\examples{
plot_worms(mse_example)
}
