% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm-functions.R
\name{P10}
\alias{P10}
\alias{P10_yrs6_20}
\alias{P10_yrs21_35}
\alias{P10_yrs36_50}
\alias{P40}
\alias{P40_yrs6_20}
\alias{P40_yrs21_35}
\alias{P40_yrs36_50}
\alias{P80}
\alias{P80_yrs6_20}
\alias{P80_yrs21_35}
\alias{P80_yrs36_50}
\alias{P100}
\alias{P100_yrs6_20}
\alias{P100_yrs21_35}
\alias{P100_yrs36_50}
\alias{PNOF}
\alias{PNOF_yrs6_20}
\alias{PNOF_yrs21_35}
\alias{PNOF_yrs36_50}
\alias{LTY}
\alias{Yield}
\alias{AAVY}
\alias{LTY_MSY}
\alias{STY_MSY}
\alias{Decline}
\title{Various performance metrics}
\usage{
P10(mse_obj)

P10_yrs6_20(mse_obj)

P10_yrs21_35(mse_obj)

P10_yrs36_50(mse_obj)

P40(mse_obj)

P40_yrs6_20(mse_obj)

P40_yrs21_35(mse_obj)

P40_yrs36_50(mse_obj)

P80(mse_obj)

P80_yrs6_20(mse_obj)

P80_yrs21_35(mse_obj)

P80_yrs36_50(mse_obj)

P100(mse_obj)

P100_yrs6_20(mse_obj)

P100_yrs21_35(mse_obj)

P100_yrs36_50(mse_obj)

PNOF(mse_obj)

PNOF_yrs6_20(mse_obj)

PNOF_yrs21_35(mse_obj)

PNOF_yrs36_50(mse_obj)

LTY(mse_obj)

Yield(mse_obj)

AAVY(mse_obj)

LTY_MSY(mse_obj, ref = 0.5, yrs = c(36, 50))

STY_MSY(mse_obj, ref = 0.5, yrs = c(6, 20))

Decline(mse_obj, ref = 0, yrs = c(6, 20))
}
\arguments{
\item{mse_obj}{An object of class \linkS4class{MSE} from the \link{DLMtool} package}
}
\value{
An object of class \linkS4class{PMobj} from the \link{DLMtool} package
}
\description{
Various performance metrics
}
\examples{
P10(mse)
PNOF_yrs6_20(mse)
}
