% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{reduce_survey}
\alias{reduce_survey}
\title{Eliminate survey data for selected years}
\usage{
reduce_survey(
  mp,
  slots = c("Ind", "VInd", "SpInd", "AddInd", "CAA", "CAL", "ML"),
  index = function(x) seq(1, x, by = 2)
)
}
\arguments{
\item{mp}{An existing management procedure function of class \code{"MP"} that will
work with \pkg{DLMtool}.}

\item{slots}{The slots for which you want to reduce observations.}

\item{index}{A function that takes the number of years of observations and
returns a vector indexing the years that should be turned into \code{NA} values.
The default anonymous function discards observations for odd years
starting in the first projection year. To start in the second projection
year, switch to \code{function(x) seq(2, x, by = 2)}.}
}
\value{
A management procedure function of class \code{"MP"} for use with
\pkg{DLMtool}.
}
\description{
This function factory creates a \pkg{DLMtool} management procedure (MP)
function that only sees survey data from selected years. For example, this
could change the data that an MP sees from an annual survey into a biennial
or triennial survey. It could also be used to eliminate commercial
observations from some years if the default slots affected were changed.
}
\examples{
library(DLMtool)
om <- DLMtool::testOM
om@nsim <- 3
temp_mp <- reduce_survey(Islope1)
# mse <- runMSE(OM = om, MPs = "temp_mp")
}
