% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_AddIerr.R
\name{sample_AddIerr}
\alias{sample_AddIerr}
\title{Sample additional index error}
\usage{
sample_AddIerr(om, beta = 1)
}
\arguments{
\item{om}{A DLMtool operating model that has been run through
\code{\link[MSEtool:SRA_scope]{MSEtool::SRA_scope()}}.}

\item{beta}{The hyperdepletion/hyperstability parameter. 1 indicates none.}
}
\value{
A DLMtool operating model.
}
\description{
This function adds log-normally distributed error to \code{OM@cpars$AddIerr}. It
also sets the survey hyperdepletion/hyperstability parameter to a specified
value (default 1).
}
\details{
This is useful if you want to override the default of estimating
hyperdepletion/hyperstability, observation error, and autocorrelation in the
additional "real" indices as is done in DLMtool.
}
\examples{
library(MSEtool)
om <- DLMtool::testOM
om@nsim <- 5
set.seed(1)
om <- SRA_scope(om,
  data = list(Chist = runif(10), Index = runif(10), I_sd = rep(0.1, 10))
)
om2 <- sample_AddIerr(om@OM)
matplot(t(om2@cpars$AddIerr[,1,]), type = "l", lty = 1)
}
