% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{SP_gf}
\alias{SP_gf}
\alias{SP6040_gf}
\alias{SP8040_gf}
\alias{SP4010_gf}
\alias{add_SP_prior}
\title{PBS groundfish surplus production wrapper function}
\usage{
SP_gf(
  x,
  Data,
  reps = 1,
  LRP = 0.4,
  TRP = 0.6,
  RP_type = "SSB_SSBMSY",
  start = list(r_prior = c(0.3, 0.1)),
  use_r_prior = TRUE,
  tac_max_increase = 1.2,
  tac_max_decrease = 0.5,
  tac_floor = 0.1,
  tac_increase_buffer = 1.05
)

SP6040_gf(x, Data, reps = 1, ...)

SP8040_gf(x, Data, reps = 1, ...)

SP4010_gf(x, Data, reps = 1, ...)

add_SP_prior(
  mp,
  r_prior,
  tac_max_increase = 1.2,
  tac_max_decrease = 0.5,
  tac_floor = 0.1,
  tac_increase_buffer = 1.05,
  ...
)
}
\arguments{
\item{x}{Index}

\item{Data}{Data}

\item{reps}{Reps}

\item{LRP}{Lower reference point}

\item{TRP}{Target reference point}

\item{RP_type}{Reference point type}

\item{start}{Starting list for the model}

\item{use_r_prior}{Logical but whether to use a prior}

\item{tac_max_increase}{Maximum proportional increase in TAC}

\item{tac_max_decrease}{Maximum proportional decrease in TAC}

\item{tac_floor}{Floor for TAC}

\item{tac_increase_buffer}{Proportional buffer below which TAC won't change}

\item{...}{Other args}

\item{mp}{MP to wrap}

\item{r_prior}{Mean and SD of r prior}
}
\description{
PBS groundfish surplus production wrapper function
}
\examples{
my_mp <- add_SP_prior(SP4010_gf, c(0.3, 0.05))
library(DLMtool)
om <- DLMtool::testOM
om@nsim <- 5
om@proyears <- 10
mse <- runMSE(om, MPs = "my_mp")
}
