% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{use_AddInd}
\alias{use_AddInd}
\title{Use the first AddInd slot}
\usage{
use_AddInd(mp)
}
\arguments{
\item{mp}{MP to use.}
}
\description{
This function factory modifies an MP to use the \code{AddInd} (and \code{CV_AddInd})
slots from the first "real" survey instead of \code{Ind}. This means that the
first survey in your SRA must be the survey that you want to mimic.
}
\examples{
library(MSEtool)
om <- DLMtool::testOM
om@nsim <- 10
set.seed(1)
sra <- SRA_scope(om,
  data = list(Chist = runif(10), Index = runif(10), I_sd = rep(0.1, 10)))
my_mp <- use_AddInd(Itarget_base)
mse <- runMSE(sra@OM, MPs = "Itarget_base")
}
