% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{use_AddInd}
\alias{use_AddInd}
\title{Use AddInd slot}
\usage{
use_AddInd(mp, i)
}
\arguments{
\item{mp}{MP to use.}

\item{i}{The i-th index in AddInd to assign to the Ind slot.}
}
\description{
This function factory modifies an MP to use an index from \code{AddInd} (and \code{CV_AddInd})
slots instead of \code{Ind} by generating a wrapper function that assigns AddInd to Ind before
calling the MP function.
}
\examples{
library(MSEtool)
om <- MSEtool::testOM
om@nsim <- 10
set.seed(1)
sra <- RCM(om,
  data = list(Chist = runif(10), Index = runif(10), I_sd = rep(0.1, 10)))
my_mp <- use_AddInd(Itarget_base)
mse <- runMSE(sra@OM, MPs = "Itarget_base")
}
