% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{use_AddInd}
\alias{use_AddInd}
\title{Use AddInd}
\usage{
use_AddInd(mp)
}
\arguments{
\item{mp}{MP to use}
}
\description{
Use AddInd
}
\examples{
my_mp <- use_AddInd(IDX)
library(DLMtool)
om <- DLMtool::testOM
om@nsim <- 5
om@proyears <- 10
mse <- runMSE(om, MPs = "my_mp")
}
